Lab 6
This is a submission of my Lab 6. They were pretty difficult. I had a
huge problem especially with number 5. I was able to get the
information, but it looks like I was making the products.name repeat
for every customers.name.

— I accidentally submitted it to a different account. Does it still
count as late even if I am a minute late? I apologize.