Lab 9
This is my Lab 9 submission. I'm not sure if I am missing anything. The
submission says to submit my runnable .sql script, but the assignment
only says to create a ER diagram. You also want us to convince you that
it is 3NF, but the instructions say to include multiple attributes. I'm
just confused, but want to do my best.